Delegating invocation retrieval to invocation finder
